services/application-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
prometheus-fastapi-instrumentator
httpx
tenacity
starlette~=0.46.2
pybreaker